+++ packages/amber/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.3.3-dev.6
+
+ - **FIX**: clean imports.
+ - **FIX**: missing convert  dependency.
+ - **FIX**: remove hex package depandance.
+
 ## 0.3.3-dev.5
 
  - **FIX**: clean imports.