+++ packages/objectbox/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.7-dev.6
+
+ - **FIX**: clean imports.
+
+## 0.2.7-dev.5
+
+ - **FIX**: clean imports.
+
 ## 0.2.7-dev.4
 
  - **FIX**: clean imports.